Promote arm64 macOS to tier 1
Github recently added a new macOS runner that is using the m1 CPU that
is usable for open source projects. [1] Previously rustworkx had support
for arm64 macOS at tier 4 because we were only able to cross compile for
the platform and not test the binaries. Now that we can run CI jobs on
the platform we're able to run both unit tests and test our binaries on
release. This commit adds a new set of test jobs and wheel builds that
use the macos-14 runner that mirrors the existing x86_64 macOS jobs we
have. This brings arm64 macOS to the same support level as arm64.

[1] https://github.blog/changelog/2024-01-30-github-actions-introducing-the-new-m1-macos-runner-available-to-open-source/