Improve ergonomics of `tox` lint jobs
This fixes the `lint` job for `tox>=4`, which uses `whitelist_externals`
is removed in favour of `allowlist_externals`.  This also makes the
`lint` and `black` runs skip installation of the package, which makes
them much faster to use while making changes, and these are static
analysis tools, so don't need to install.