Finish rustworkx-core port
This commit finishes the rustworkx-core port of the steiner tree and
metric closure functions. In particular this was especially tricky
because the petgraph traits make it exceedingly difficult to have a
generic function that takes in a graph for analysis and modifies it as
most of the traits required for visiting/iteration that are used for
analysis are only defined on borrowed graphs, and the limited traits for
modifying graphs are defined on owned graph types. This causes a
conflict where you can't easily express that a generic type G created in
a function from a user input is both mutated using a trait and analyzed
as there is a type mismatch between G and &G. After spending far too
long to fail to find a pattern to express this, I opted to just use a
discrete type for the return and leave the actual graph mutation up to
the rustworkx-core user because we're lacking the ability to cleanly
express what is needed via petgraph.