Fix backport branch post 0.14.0 release (#1101)
In #1070 we updated the version numbers of rustworkx to indicate the
main branch is developing the next release. However, in that PR we
neglected to update the mergify backport branch. This means when we tag
a PR to main as stable backport potential mergify will incorrectly
backport the PR to stable/0.13 instead of stable/0.14. This commit
corrects this oversight and updates the mergify config to backport to
the correct branch.

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>