Add manual iterator implementations for custom vec iterables
This adds manual `Iterator` implementors for the custom vec `Iterable`
pyclasses, which were previously using the Python default iterator
implementation that attempts calls to `__getitem__` with successive
integers until an `IndexError` is raised.

Doing this means that we no longer have any overhead from converting
Python-space integer indices to Rust-space ones, because we just store
a pointer to the backing vec internally and yield from that using an
index we track ourselves.