Promote arm64 macOS to tier 1 (#1077)
* Promote arm64 macOS to tier 1

Github recently added a new macOS runner that is using the m1 CPU that
is usable for open source projects. [1] Previously rustworkx had support
for arm64 macOS at tier 4 because we were only able to cross compile for
the platform and not test the binaries. Now that we can run CI jobs on
the platform we're able to run both unit tests and test our binaries on
release. This commit adds a new set of test jobs and wheel builds that
use the macos-14 runner that mirrors the existing x86_64 macOS jobs we
have. This brings arm64 macOS to the same support level as arm64.

[1] https://github.blog/changelog/2024-01-30-github-actions-introducing-the-new-m1-macos-runner-available-to-open-source/

* Specify patch version for main test ci job

The setup-python action on macos-14 only has python 3.12 installers for
specific patch versions. This commit specifies that we use 3.12.0 as
this lets us start using stestr again to improve test throughput. Using
3.12.1 is blocked as there was a breaking api change made to unittest in
that Python release which is causing skip tests to be reported as errors.

* Use correct architecture string for Python

* Skip py3.8 and 3.9 until the setup-python action is updated

* Revert Python 3.12 job constraint and stestr usage