Pin lxml in constraints
The recent lxml 5.2.0 release is raising an exception about a packaging
change that moved a submodule into a different location during the docs
build. We don't actually directly use lxml in rustworkx (or the docs),
but it gets pulled in via a dependency of jupyter-sphinx (it's deeper
in the dependency tree than jupyter-sphinx's immediate dependencies). To
workaround this until things are updated to take this into account this
commit pins the lxml version in the constraints.txt file so we don't
pull in the latest release.