Avoid using macos-latest for jobs we want x86_64 macOS
Github actions recently changed the macos-latest runner to use macos-14,
which changes the CI environment to run with arm64 instead of x86_64.
This has several impacts on our test jobs where the intent was to
specifically test native x86_64 macOS. This commit changes the ci
configuration to use macos-13 in places where this was the case, as we
were already explicitly using macos-14 where we wanted arm64 runners.