Add lexicographical_topological_sort function to rustworkx-core (#1197)
* Add lexicographical_topological_sort function to rustworkx-core

This commit adds a rust generic implementation of the lexicographical
topological sort function that was previously only exposed via the
python interface to rustworkx-core. This new function is used internally
by rustworkx and replaces the python specific implementation that used
to be there. But it now exposes an interface to use it to rust.

Fixes: #1165

* Simplify trait bounds

* Use .lcov extension for coveralls

---------

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>