Use stestr again for unittests (#1218)
With the release of Python 3.12.1 we stopped using stest in CI because
3.12.1 introduced a breaking API change to the unittest runner API that
broke a dependency of stestr, testtools. This has been worked around
with a new testtools release so we can restore using stestr in CI again
which should improve test run throughput. This commit restores using
stestr for the test CI jobs and local runs. The cibuildwheel config
remains unchanged and uses the stdlib unittest runner still, just for
simplicity so we don't need to install additional test dependencies.