Prepare 0.15.0 release
This commit prepares the 0.15.0 release. It moves the release notes for
the release into a separate directory and adds a prelude to the release
notes. This should be the last commit to merge for the 0.15.0 release
and after it merges that commit will be what is should be tagged as the
0.15.0 release.

As part of this there are few small changes other changes being made.
The release notes are update to have consistent formatting and wording.