Prepare 0.15.0 release (#1228)
* Prepare 0.15.0 release

This commit prepares the 0.15.0 release. It moves the release notes for
the release into a separate directory and adds a prelude to the release
notes. This should be the last commit to merge for the 0.15.0 release
and after it merges that commit will be what is should be tagged as the
0.15.0 release.

As part of this there are few small changes other changes being made.
The release notes are update to have consistent formatting and wording.

* Add missing release notes

* Apply suggestions from code review

Co-authored-by: Raynel Sanchez <87539502+raynelfss@users.noreply.github.com>

* Move new release note

* Reword Dorogovtsev-Goltsev-Mendes rustworkx-core release note slightly

* Update releasenotes/notes/0.15/swap-nox-tox-dea2bb14c400641c.yaml

* Update releasenotes/notes/0.15/prepare-0.15.0-771047bc5b71c569.yaml

---------

Co-authored-by: Raynel Sanchez <87539502+raynelfss@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>