Fix dot file generation
This commit fixes a regression introduced in #1203. That PR was
attempting to fix missing character escaping in some string fields
but it was doing so too eagerly. This would result in invalid dot
files being produced for users upgrading from rustworkx < 0.15.0 that
were wrapping strings in quotes as needed previously. For example if you
were setting `'color="#aaaaaa"'` previously before this would become
`color="\"#aaaaaa\""` after #1203. In order to quickly release a 0.15.1
this commit reverts the dot generation component of #1203 but updates
the code to wrap tooltip in addition to label which was what the
original bug reported. In 0.16.0 we should investigate adding a flag to
control the escaping behavior of the function to either decide to wrap
values in quotes or not.