Advanced Tests for rust repo (#1403)
* Advanced Tests for rust repo

Starting off with the implementation of cargo-fuzz which is used to find panics.

Implemented the "test_is_connected" test in the fuzz. Working on more.

* Removing log statement

* Reverting the changes in dependencies.

* Updating contributing.md file

The fuzz tests are build with the command cargo build.
The steps to run a particular tests, or add new ones are added in the Contributing.md file.

* Fixed lint issues

* Updates based on feedback

1) Created a new test to assert whether the nodes visited by dfs and bfs are same.

2) Its not possible to run all the fuzz targets at once, so added another comment in the contributing file that will list all the fuzz targets.

* Updating the contributing.md file

---------

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>