Bump petgraph to 0.8.1 (#1424)
This commit bumps the petgraph version used in rustworkx to the latest
release 0.8.1. This is technically an api change for rustworkx-core as
some of the generics (mostly around MatrixGraph) changed, but also
because now when building rustworkx-core only petgraph 0.8.1 objects are
accepted and older releases are no longer compatible with
rustworkx-core. This is not generally an issue if you use the
rustworkx-core petgraph re-export at ``rustworkx_core::petgraph``, but
as 0.8.1 does contain some api changes users will need to account for
those when upgrading to rustworkx-core 0.17.0.