Add distance_matrix to rustworkx-core (#1439)
* Add distance_matrix to rustworkx-core

This commit moves the distance matrix functionality to rustworkx. This
is mostly a straightforward migration as the functionality was written
in a generic way already. The only difference is how node holes were
handled, the this opted to split the functions into 2, a verison that
assumes the graph has compact indices and one that doesn't.

* Optimize the implementation of distance matrix

This commit performs some optimizations on the internals of the
distance_matrix() function. It avoids extra allocations and uses a
fixedbitset for tracking instead of hashsets.

Co-authored-by: Jake Lishman <jake@binhbar.com>

* Deduplicate functions

* Add release note

* Remove stray debug print

* Remove invalid classifier and capitalize keywords

* Bump release version to 0.17.1

---------

Co-authored-by: Jake Lishman <jake@binhbar.com>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <ivancarvalho@gatech.edu>