Fix `pyproject.toml` and stop using `setup.py` (#1494)
* Add distance_matrix to rustworkx-core (#1439) (#1493)

* Add distance_matrix to rustworkx-core

This commit moves the distance matrix functionality to rustworkx. This
is mostly a straightforward migration as the functionality was written
in a generic way already. The only difference is how node holes were
handled, the this opted to split the functions into 2, a verison that
assumes the graph has compact indices and one that doesn't.

* Optimize the implementation of distance matrix

This commit performs some optimizations on the internals of the
distance_matrix() function. It avoids extra allocations and uses a
fixedbitset for tracking instead of hashsets.



* Deduplicate functions

* Add release note

* Remove stray debug print

* Remove invalid classifier and capitalize keywords

* Bump release version to 0.17.1

---------




(cherry picked from commit 7318a803dc87853cbb2e916d7095b440985507e0)

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Jake Lishman <jake@binhbar.com>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <ivancarvalho@gatech.edu>

* Make workflow only trigger sdist steps

* Tweak license field

* Fix retworkx dist path

* Limit trigger to today

* Update timestamp condition in wheels.yml

* Update build conditions for wheels on specific date

* Revert to main status

* Use build for wheels.yml

* Fix newline

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Jake Lishman <jake@binhbar.com>